# internal/handlers/farm_handlers.go
package handlers

import (
	"net/http"

	"github.com/Kisanlink/farmers-module/internal/services"
	"github.com/gin-gonic/gin"
)

// CreateFarm handles W6: Create farm
// @Summary Create a new farm
// @Description Create a new farm with geographic boundaries and metadata
// @Tags farms
// @Accept json
// @Produce json
// @Param farm body object true "Farm data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /farms [post]
func CreateFarm(service services.FarmService) gin.HandlerFunc {
	return func(c *gin.Context) {
		var req struct {
			FarmerID string `json:"farmer_id" binding:"required"`
			FPOID    string `json:"fpo_id" binding:"required"`
			Name     string `json:"name" binding:"required"`
			Geometry string `json:"geometry" binding:"required"`
			Location string `json:"location" binding:"required"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm created successfully",
			"data": gin.H{
				"farmer_id": req.FarmerID,
				"fpo_id":    req.FPOID,
				"name":      req.Name,
				"geometry":  req.Geometry,
				"location":  req.Location,
			},
		})
	}
}

// UpdateFarm handles W7: Update farm
// @Summary Update an existing farm
// @Description Update farm details including name, geometry, and location
// @Tags farms
// @Accept json
// @Produce json
// @Param farm_id path string true "Farm ID"
// @Param farm body object true "Farm update data"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /farms/{farm_id} [put]
func UpdateFarm(service services.FarmService) gin.HandlerFunc {
	return func(c *gin.Context) {
		farmID := c.Param("farm_id")
		var req struct {
			Name     *string `json:"name,omitempty"`
			Geometry *string `json:"geometry,omitempty"`
			Location *string `json:"location,omitempty"`
		}

		if err := c.ShouldBindJSON(&req); err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm updated successfully",
			"data": gin.H{
				"farm_id": farmID,
			},
		})
	}
}

// DeleteFarm handles W8: Delete farm
// @Summary Delete a farm
// @Description Delete a farm by ID
// @Tags farms
// @Accept json
// @Produce json
// @Param farm_id path string true "Farm ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /farms/{farm_id} [delete]
func DeleteFarm(service services.FarmService) gin.HandlerFunc {
	return func(c *gin.Context) {
		farmID := c.Param("farm_id")

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm deleted successfully",
			"data": gin.H{
				"farm_id": farmID,
			},
		})
	}
}

// ListFarms handles W9: List farms
// @Summary List all farms
// @Description Retrieve a list of all farms with optional filtering
// @Tags farms
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /farms [get]
func ListFarms(service services.FarmService) gin.HandlerFunc {
	return func(c *gin.Context) {
		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farms retrieved successfully",
			"data":    []interface{}{},
		})
	}
}

// GetFarm handles getting farm by ID
// @Summary Get farm by ID
// @Description Retrieve a specific farm by its ID
// @Tags farms
// @Accept json
// @Produce json
// @Param farm_id path string true "Farm ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]interface{}
// @Router /farms/{farm_id} [get]
func GetFarm(service services.FarmService) gin.HandlerFunc {
	return func(c *gin.Context) {
		farmID := c.Param("farm_id")

		// TODO: Implement the actual service call
		c.JSON(http.StatusOK, gin.H{
			"message": "Farm retrieved successfully",
			"data": gin.H{
				"farm_id": farmID,
			},
		})
	}
}